Partial fix to versioning problems
The cmdlets use the 18.1 API by default, and few use a some newer APIs,
but they do not use the right API for the right server. This leads to
serialization issues like #112 or #125 when newer Orchestrator responds
with new values for old enums and the deserialization fails.

This should unblock GetMachine, GetUser and getRobot on UiPath Platform
(API v11 (20.4)) but a more thorough solution is needed long term to
have each API call be routed through the API client that the server
expects.

Fixes #112
Fixes #125